// src/app/model/customers.ts
export interface Customer {
    first_name: string;
    last_name: string;
      email: string;
      phone_number: string;
      address: string;
      loyalty_points: number;
      discount_eligibility: boolean;
      created_at: Date;
      updated_at: Date;
      biographic_info: string;
    }
    